indeling componenten
