+++ TG-Netflix/src/components/account-page/SettingsSection.jsx
@@ -1,0 +1,32 @@
+import React from 'react'
+
+export default function SettingsSection() {
+  return (
+    <div className="account-section">
+          <div className="left-side">
+            <h2>SETTINGS</h2>
+          </div>
+          <div className="account-subsection settings">
+            <div className="links">
+              <ul>
+                <li>
+                  <a href="">Test participation</a>
+                </li>
+                <li>
+                  <a href="">Manage download devices</a>
+                </li>
+                <li>
+                  <a href="">Recent device streaming activity</a>
+                </li>
+                <li>
+                  <a href="">Sign out of all devices</a>
+                </li>
+                <li>
+                  <a href="">Download your personal information</a>
+                </li>
+              </ul>
+            </div>
+          </div>
+        </div>
+  )
+}
